Refactor plugin import process
