+++ plugins/plugin/plugin.go
@@ -1,0 +1,13 @@
+package plugin
+
+import goship "github.com/gengo/goship/lib"
+
+var Plugins []Plugin
+
+type Plugin interface {
+	Apply(goship.Config) error
+}
+
+func RegisterPlugin(p Plugin) {
+	Plugins = append(Plugins, p)
+}
