+++ plugins/plugins.go
@@ -1,13 +1,5 @@
 package plugins
 
-import goship "github.com/gengo/goship/lib"
-
-var Plugins []Plugin
-
-type Plugin interface {
-	Apply(goship.Config) error
-}
-
-func RegisterPlugin(p Plugin) {
-	Plugins = append(Plugins, p)
-}
+import (
+	_ "github.com/gengo/goship/plugins/helloworld"
+)
